[ChannelOut] Skip frames when we are speeding up too fast.
This proof of concept patch will skip frames if the sequence
player is running more than 2ms faster than the sequence file
specifies because it is trying to catch up to the media.

This is to test whether we are sending data too fast out the
SPI port to the FPD.  This patch affects all channel outputs
though, not just the FPD, so it is only a proof of concept to
see if the issue is what we think it is.